requirements.txt
keras==2.12.0
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
pywhatkit==5.3
requests==2.31.0
scikit_learn==1.2.2
python-dotenv==1.0.0
tensorflow==2.12.0
